Moving task-event-repeater from container to module
