--- blocks/core/ff_container/ff_container-task-event-repeater/ff_container-task-event-repeater.md
@@ -1,1 +1,0 @@
-<div data-ff-container-task-event-repeater=""/>
