+++ blocks/core/ff_module/ff_module-task-event-repeater/ff_module-task-event-repeater.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var React = require('react');
-var template = require('./_ff_container-task-event-repeater.rt.js');
+var template = require('./_ff_module-task-event-repeater.rt.js');
 
 module.exports = React.createClass({
     render: template
